Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,10 +1,10 @@
 #include <stdio.h>
 int main() {
-    double a,b,c;
+    int a,b,c;
     scanf("%d\t",&a);
     scanf("%d\t",&b);
     scanf("%d\t",&c);
-    float d=(a+b+c)/3.0;
+    double d=(a+b+c)/3.0;
     printf("Average: %.2lf\n",d);
     return 0;
 }
